Move chat-components into chat-folder
