+++ src/lib/components/UI.svelte
@@ -2,9 +2,9 @@
 	import Zoom from '$lib/components/Tools/Zoom.svelte';
 	import ServerId from './ServerId.svelte';
 	import Menu from './Menu.svelte';
-	import ChatButton from './ChatButton.svelte';
+	import ChatButton from './Chat/ChatButton.svelte';
 	import Navbar from './Navbar/Navbar.svelte';
-	import Chat from './Chat.svelte';
+	import Chat from './Chat/Chat.svelte';
 
 	import { viewChat } from '$lib/stores/stateStore';
 	import SettingsModal from './SettingsModal.svelte';
